fix: Check the error return value of p.StartHTTP and fix ineffassign in ca package
Error return value of the p.StartHTTP was not checked in main. Also, in the ca package (ca/ca.go) there was an ineffectual assignment (ineffassign) to the ca variable, as even thought it was checked for Leaf, it was parsed right away in the next line. Now it uses the else block for parsing.