fix: Filtering wasn't applied due to port being present in the request host
The host and port split was not implemented in the ApplyFilter() function. Because of this, the filtering saw host:port, and the blocklist wasn't applied. Moved up the host and port split to ServeHTTP() level, so it is now implemented for each handler function, removing duplicates and also making it available for the filtering function.