fix: Move back port strip to functions level within handler
Since handleConnect the forwardRequest methods within the handler build up new connections to the target, they need to use the port number as well. This means that, stripping the port on handler level is not a good idea, as the proxy will fail to build up connections to the target server. To fix this, the port strips have been moved down to functions level.