fix: Runtime panic and use any type in runPlugins
The code tried to reach an error when writing to the client, however, there was no error as it was a Cancel outcome without error. Also, temporarily changedd the input parameter for runPlugins to accept a single any parameter that is type asserted to *http.Request, and will be type asserted to *http.Response when working with responses in any of the upcoming plugins.