test: Log client request data after `CONNECT` is completed
As described in https://github.com/akos011221/armor/issues/27, currently it is not possible to inspect data after the `CONNECT` method is completed, so the plugins are not seeing the real client<->server communication. This test uses TeeReader, so the client request can be logged without actually consuming the data (i.e: we can still forward it to the target server).