test: Replace Writer in TeeReader()
The current solution is wrong, as tee.Read() writes to the actual network connection and then io.Copy() also tries to writes to the same, but probably nothing left for it. The TeeReader was not used properly. Now, the TeeReader writes to the proxy's log destination and then io.Copy can use its returned Reader to write the data to the target TLS connection.