feat: Mitm processing for plugins
Implement the plugin processing in the MITM connection (i.e: the connection between the client and server after hijack). A seperate function in the plugin plackage is implemented; plugins can have processing for _CONNECT_ requests and for the actual _payload_. The response processing was removed for now, as it is not used.